requirements.txt
python-dotenv
transformers
sentence-transformers
langchain==0.1.0
langchain-openai
